LPS/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
Django==4.0
djangorestframework==3.12.4
ecdsa==0.17.0
mysqlclient==2.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.19.2
python-jose==3.3.0
pytz==2021.3
rsa==4.8
six==1.16.0
sqlparse==0.4.2
psycopg2-binary

